feat(cron-jobs): refactored env functions and align cron.js layout to the provided example, added cron execution in db index as per pr feedback
